third_party/WebKit/LayoutTests/inspector/elements/elements-css-path-expected.txt

Tests DOMNode.cssPath()

ид
класс
   html
    head
      head > meta
      #script-id
      head > script:nth-child(3)
      #test-script
    body
      body > p
      body > article:nth-child(2)
      body > article:nth-child(3)
      #ids
        #ids > div:nth-child(1)
        #ids > div:nth-child(2)
        #inner-id
        #__proto__
        #\23 \22 ridiculous\22 \2e id
        #\27 quoted\2e value\27
        #\2e foo\2e bar
        #\2d
        #-a
        #\2d 0
        #\37
        #ид
        #ids > p
      #classes
        #classes > div.foo.bar
        #classes > div:nth-child(2)
        #classes > div.\2e foo
        #classes > div.\2e foo\2e bar
        #classes > div.\2d
        #classes > div.-a
        #classes > div.\2d 0
        #classes > div.\37
        #classes > div.класс
        #classes > div:nth-child(10)
        #classes > div.__proto__.foo
        #classes > span
        #id-with-class
      #non-unique-classes
        #non-unique-classes > span:nth-child(1)
        #non-unique-classes > span:nth-child(2)
        #non-unique-classes > span:nth-child(3)
        #non-unique-classes > span.c1.c2.c3
        #non-unique-classes > span:nth-child(5)
        #non-unique-classes > div:nth-child(6)
        #non-unique-classes > div.c1.c2
        #non-unique-classes > div.c3.c2
        #non-unique-classes > div.c3.c4
        #non-unique-classes > div.c1.c4
        #non-unique-classes > div:nth-child(11)
